--- backend/js/src/main/scala/zio/logging/backend/PlatformSpecificBackends.scala
@@ -1,3 +1,0 @@
-package zio.logging.backend
-
-trait PlatformSpecificBackends {}
