--- slf4j-bridge/src/main/scala/zio/logging/slf4j/bridge/package.scala
@@ -1,18 +1,0 @@
-package zio.logging.slf4j
-
-import org.slf4j.impl.ZioLoggerFactory
-import zio.logging.Logging
-import zio.{ Tag, ZIO, ZLayer }
-
-package object bridge {
-  def initializeSlf4jBridge[R <: Logging: Tag]: ZLayer[R, Nothing, R] =
-    ZIO
-      .runtime[R]
-      .flatMap { runtime =>
-        ZIO.succeed {
-          ZioLoggerFactory.initialize(runtime)
-          runtime.environment
-        }
-      }
-      .toLayerMany
-}
